+++ include/string.h
@@ -4,15 +4,15 @@
 
 #include <stddef.h>
 
-int strlen(const char *str);
+unsigned long strlen(const char *str);
 char *strcpy(char *dest, const char *src);
-char *strncpy(char *dest, const char *src, int max);
+char *strncpy(char *dest, const char *src, unsigned long max);
 int strcmp(const char *str1, const char *str2);
-int strncmp(const char *str1, const char *str2, int max);
+int strncmp(const char *str1, const char *str2, unsigned long max);
 const char *strchr(const char *str, int ch);
 const char *strrchr(const char *str, int ch);
 char *strcat(char *dest, const char *src);
-char *strncat(char *dest, const char *src, int max);
+char *strncat(char *dest, const char *src, unsigned long max);
 size_t strspn(const char *str1, const char *str2);
 size_t strcspn(const char *str1, const char *str2);
 char *strstr(const char *str1, const char *str2);
